% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_duration.R
\name{as_duration}
\alias{as_duration}
\title{Safe Version of as.duration}
\usage{
as_duration(...)
}
\description{
a version of lubridate::as.duration that returns NA on errors instead of throwing an error
}
