% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcalc.R
\name{distcalc}
\alias{distcalc}
\title{Calculate Euclidian Distance}
\usage{
distcalc(lon, lat, y)
}
\arguments{
\item{lon}{a vector of longitudues}

\item{lat}{a vector of latitudes}

\item{y}{a single destinatoin}
}
\value{
a numeric vector of distances
}
\description{
calculate Euclidean distance, a rough proxy for as-the-crow flies distance between vectors of latitude and longitudes and a single destination. Best used with mutate().
}
