% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivablecountiesvsregion.R
\name{drivablecountiesvsregion}
\alias{drivablecountiesvsregion}
\title{Compare an Existing Region to a List of Drivable Counties}
\usage{
drivablecountiesvsregion(regiondf, drivingdistance, countymaps = NA,
  fipscodes = NA, output = c("nearbydf", "mapdata", "outsidecountiesdf",
  "map"))
}
\arguments{
\item{regiondf}{an existing data frame or a file URL}

\item{drivingdistance}{an object of class drivingdistance or a numeric distance}

\item{countymaps}{a spatialpolygonsdataframe. If NA, data is downloaded and read from the US Census}

\item{fipscodes}{a data frame of fips codes for all counties with bannerized codes added. If NA, data from the tigris package is modified to work}

\item{output}{a character vector describing what output you want. Available options include, "nearbydf", "mapdata", "outsidecountiesdf" and "map"}
}
\value{
if more than one output is selected, a list containing values of each. If only one output is selected, the object itself is return, without being wrapped in a list. Outputs include:
\itemize{
 \item{"nearbydf"}{A data frame of nearby counties and their distance to centerlocation}
 \item{"mapdata"}{A SpaitalPolygonsDataFrame of the nearby counties appropriate for use in a leaflet map}
 \item{"outsidecountiesdf"}{A data frame of counties that are within driving range but outside the regiondf}
 \item{"map"}{A pre-generated leaflet map}
 }
}
\description{

}

