% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcenterofpolygons.R
\name{getcenterofpolygons}
\alias{getcenterofpolygons}
\title{Get the Center of a Bunch of Polygons}
\usage{
getcenterofpolygons(polygons)
}
\arguments{
\item{polygons}{a SpatialPolygonsDataFrame}
}
\value{
a two column data frame with the longitude and latitude of the center point of polygons
}
\description{
Get the Center of a Bunch of Polygons
}
