% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcountymaps.R
\name{getcountymaps}
\alias{getcountymaps}
\title{Get County Map}
\usage{

  getcountymaps(filelocation = "http://www2.census.gov/geo/tiger/GENZ2015/shp/cb_2015_us_county_20m.zip",
  dest = "data", local = F)
}
\arguments{
\item{filelocation}{a character string describing where to get the shape files. By default, assumes the files should be downloaed from the US Census Bureau at https://www.census.gov/geo/maps-data/data/cbf/cbf_counties.html. Local file addresses can also be used.}

\item{dest}{the local destination where the files should be downloaded to.}

\item{local}{a logical value indicating whether filelocation is a local address or not.}
}
\value{
a SpatialPolygonsDataFrame with polygons for each county within the United States.
}
\description{
Create a SpatialPolygonsDataFrame of all the counties in the continental US
}

