% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdrivingtime.R
\name{getdrivingtime}
\alias{getdrivingtime}
\title{Get Driving Time from Google Query}
\usage{
getdrivingtime(results)
}
\arguments{
\item{results}{the results of a `googleway::google_distance()` query}
}
\value{
a vector of driving times with class `lubridate::duration`. When the google query returns no results, driving time is returned as NA.
}
\description{
Get Driving Time from Google Query
}
