% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_distance_loud.R
\name{google_distance_loud}
\alias{google_distance_loud}
\title{Get Driving Distance to/from a place}
\usage{
google_distance_loud(x, y, thisdest = placeofinterest$placename, thiskey = key)
}
\arguments{
\item{x}{a one row, two column data_frame with lat and lon columns in that order}

\item{y}{a indicator of what county you're getting distance to/from}

\item{thisdest}{the destination}

\item{thiskey}{a Google Maps API key}
}
\description{
a version of googleway::google_distance that messages where you're at. Used in countieswithindrivingdistance(), not exported.
}
