% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebannerizedfips.R
\name{makebannerizedfips}
\alias{makebannerizedfips}
\title{Bannerize FIPS Codes}
\usage{
makebannerizedfips()
}
\value{
a data frame of all the FIPS codes in the US.
}
\description{
modify the default list of fips county codes to add columns for joining with Banner and Census Bureau. Uses the fips_codes data from the tigris package.
}
