% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peoplenearbydist.R
\name{peoplenearbydist}
\alias{peoplenearbydist}
\title{People with a Given Distance}
\usage{
peoplenearbydist(locations, center, maxdist, key)
}
\arguments{
\item{locations}{a data frame that includes lat and lon columns}

\item{center}{a center location. Can be a character string or a two column data frame.}

\item{maxdist}{maximum distance in miles (converted to Euclidean distance by rough calculation)}

\item{key}{Google API key (see https://developers.google.com/maps/documentation/geocoding/get-api-key)}
}
\value{
a data frame with mapdist columns appended
}
\description{
Given a data frame of locations with lon and lat columns and a destination, filter the data frame to return only results that are within maxdist of the destination.
}
