% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peoplewithindrivingdistance.R
\name{peoplewithindrivingdistance}
\alias{peoplewithindrivingdistance}
\title{People Within Driving Distance}
\usage{
peoplewithindrivingdistance(locations, center, maxtime, key)
}
\arguments{
\item{locations}{a data frame of locations including  lon and lat columns}

\item{center}{a center location. Can be a character string or a two column data frame.}

\item{maxtime}{a maximum driving time in minutes}

\item{key}{Google API key (see https://developers.google.com/maps/documentation/geocoding/get-api-key)}
}
\value{
a filtered data frame, with only rows where location is with maxtime minutes of center
}
\description{
People Within Driving Distance
}
