% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapregion.R
\name{mapregion}
\alias{mapregion}
\title{Map a Geo Region}
\usage{
mapregion(regionfile, countymaps = NA, fipscodes = NA)
}
\arguments{
\item{regionfile}{the location of a csv with one column: bannercode}

\item{countymaps}{a spatialpolygonsdataframe. If NA, data is downloaded and read from the US Census}

\item{fipscodes}{a data frame of fips codes for all counties with bannerized codes added. If NA, data from the tigris package is modified to work}
}
\value{
a spatialPolygonsDataFrame of the region. As a side effect, generates a leaflet map of the counties
}
\description{
Map a Geo Region
}

